requirements.txt
asgiref==3.8.1
Django==3.2.18
djangorestframework==3.15.1
pytz==2024.1
sqlparse==0.5.0
